- Unify type of contents to dict[str, any] in both JSON and non-JSON cases
